// 3.Sourcecode/front-end/ecommerce/src/app/admin/quan-ly-san-pham/model/product-admin-response.model.ts
export interface ProductAdminResponse{
      id: number;
      productName : string;
      shortDesciption: string;
      cost: number;
      quantity: number;
      categoryId: number;
      colorId: number;
      sizeId: number;
      discountId: number;
      createdDate: number;
      updatedDate: number;
}